requirements.txt
aerich==0.7.2
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
cryptography==42.0.2
Cython==3.0.8
dictdiffer==0.9.0
distro==1.9.0
frozendict==2.4.0
frozenlist==1.4.1
future==0.18.3
genanki==0.13.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iso8601==1.1.0
mikatools==1.0.1
multidict==6.0.5
numpy==1.26.4
openai==1.12.0
pandas==2.2.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pyhfst==1.1.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
tenacity==8.2.3
tinydb==4.8.0
tomlkit==0.12.3
tortoise-orm==0.20.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
uralicNLP==1.5.1
urllib3==2.2.0
wheel==0.42.0
yarl==1.9.4
